SIM Fixed up!
So it turns out using the physics array has issues because it includes the sun, which just pulls the spacecraft directly into it. If I had to guess why I would say it's because the sphere static-mesh provided in the example content has a default radius of 50. This means all our radii get multiplied by 50 when put into use. I'll tune this up to see what happens and try and fix it.

Mission can now click and drag on spacecraft to move, it's planar for now, but I'll make it work against the viewport eventually.

Minor fixes to the system level.