Even more math!
Fixed bug in sim where integrator used initial planet locations for distance measuring.
Updated displays to count properly.
Planets move!!
TODO: kepler's third law for the period so you can find out how fast the planets are rotating.
shorthand: you need the time period (third law) of the orbit to calculate the angular velocity of the planet.
sure you could hard code it to be something believable, but then the moon wouldn't work.
and it all falls apart.
don't forget the end goal of hohmann transfer by the way. so far all your code is built to handle the initial conditions, and nothing is set for the runtime. This will be a problem. Solve it.