The math the previous push refers to.
A whole bunch of fixes, including for that which isn't used, but might be. See - physics base's semi-implicit Euler integrator. I really would like it to get used, so they move smoothly. But I don't expect to be able to pull off that many miracles. We'll see. Good luck, me.